+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-19 23:29:11
-// Commit: fix(ui): prevent double tap duplicate action in ui
+// Updated at: 2026-06-21 15:08:35
+// Commit: feat(ui): add async ui stream provider
 
 class uiComponentState {
-  final String commitRef = '2166';
-  final String lastMessage = 'fix(ui): prevent double tap duplicate action in ui';
-  final DateTime timestamp = DateTime.parse('2026-06-19T23:29:11+05:00');
+  final String commitRef = '2172';
+  final String lastMessage = 'feat(ui): add async ui stream provider';
+  final DateTime timestamp = DateTime.parse('2026-06-21T15:08:35+05:00');
 }