+++ CHANGELOG.md
@@ -2173,3 +2173,4 @@
 - [2026-06-21 10:55] feat(test): integrate custom telemetry logger in test
 - [2026-06-21 13:27] feat(analytics): integrate custom telemetry logger in analytics
 - [2026-06-21 15:04] chore(theme): update dependencies and cleanup unused imports in theme
+- [2026-06-21 15:08] feat(ui): add async ui stream provider